Revert "Feature front letter refactoring"
