+++ client/src/pages/Friend/Letter.jsx
@@ -1,14 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import { formatDate } from '../utils';
+import { formatDate } from './utils';
 import { Link, useParams } from 'react-router-dom';
 import axios from 'axios';
-import LetterWrapper from '../LetterList/LetterListStyle';
+import { LetterWrapper } from './LetterStyle';
 import style from './LetterDetail.module.css';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import { useSelector } from 'react-redux';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 
-const LetterDetail = () => {
+const Letter = () => {
   const friendId = useParams().id;
   const postId = useParams().postId;
   const token = useSelector(state => state.auth.token);
@@ -31,16 +30,6 @@ const LetterDetail = () => {
     }
   };
 
-  const handleCopyClipBoard = async text => {
-    try {
-      await navigator.clipboard.writeText(text);
-
-      alert('복사 성공!');
-    } catch (error) {
-      alert('복사 실패!');
-    }
-  };
-
   useEffect(() => {
     fetchLetter();
   }, []);
@@ -52,21 +41,12 @@ const LetterDetail = () => {
       </Link>
 
       <li className={style.letterContainer}>
-        <div className={style.letterHeader}>
-          <span
-            className={style.letterCopyIcon}
-            title="복사"
-            onClick={() => handleCopyClipBoard(letter.content)}
-          >
-            <ContentCopyIcon />
-          </span>
-        </div>
         <pre className={style.letterContent}>{letter.content}</pre>
         <p className={style.letterSender}>{letter.nickname}</p>
-        <p className={style.receiveDate}>{formatDate(letter.receiveDate)}</p>
+        <p className={style.receiveDate}>{formatDate(letter.receive_date)}</p>
       </li>
     </LetterWrapper>
   );
 };
 
-export default LetterDetail;
+export default Letter;
