--- client/src/pages/Friend/LetterList/LetterListStyle.jsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const LetterWrapper = styled.ul`
-  overflow: hidden;
-  position: relative;
-  margin: 15rem 1rem 20rem 1rem;
-  display: flex;
-  flex-wrap: wrap;
-  width: 90%;
-  gap: 1.5rem;
-  text-decoration: none;
-`;
-
-export default LetterWrapper;
